Fix escaped characters on calendar import & export #4685 #3507 #2164
