+++ src/calendar/export/CalendarImporterDialog.ts
@@ -6,7 +6,8 @@ import { showProgressDialog } from "../../gui/dialogs/ProgressDialog"
 import { ParserError } from "../../misc/parsing/ParserCombinator"
 import { Dialog } from "../../gui/base/Dialog"
 import { lang } from "../../misc/LanguageViewModel"
-import { parseCalendarFile, ParsedEvent, serializeCalendar } from "./CalendarImporter"
+import { serializeCalendar } from "./CalendarExporter.js"
+import { parseCalendarFile, ParsedEvent } from "./CalendarImporter.js"
 import { elementIdPart, isSameId, listIdPart } from "../../api/common/utils/EntityUtils"
 import type { UserAlarmInfo } from "../../api/entities/sys/TypeRefs.js"
 import { createDateWrapper, UserAlarmInfoTypeRef } from "../../api/entities/sys/TypeRefs.js"
