+++ src/calendar/view/eventpopup/CalendarEventPopup.ts
@@ -14,6 +14,7 @@ import { prepareCalendarDescription } from "../../date/CalendarUtils.js"
 import { BootIcons } from "../../../gui/base/icons/BootIcons.js"
 import { IconButton } from "../../../gui/base/IconButton.js"
 import { CalendarEventPopupViewModel } from "./CalendarEventPopupViewModel.js"
+import { convertTextToHtml } from "../../../misc/Formatter.js"
 
 /**
  * small modal displaying all relevant information about an event in a compact fashion. offers limited editing capabilities to participants in the
@@ -40,7 +41,7 @@ export class CalendarEventPopup implements ModalComponent {
 		this.sanitizedDescription = prepareCalendarDescription(
 			model.calendarEvent.description,
 			(s) =>
-				htmlSanitizer.sanitizeHTML(s, {
+				htmlSanitizer.sanitizeHTML(convertTextToHtml(s), {
 					blockExternalContent: true,
 				}).html,
 		)
