+++ src/misc/Formatter.ts
@@ -143,6 +143,10 @@ export function urlEncodeHtmlTags(text: string): string {
 	return text.replace(/&/g, "&amp;").replace(/</g, "&lt;").replace(/>/g, "&gt;").replace(/"/g, "&quot;").replace(/'/g, "&#039;")
 }
 
+export function convertTextToHtml(text: string) {
+	return text.replace(/(\r)?\n/g, "<br>")
+}
+
 export function getHourCycle(userSettings: UserSettingsGroupRoot): "h12" | "h23" {
 	return userSettings.timeFormat === TimeFormat.TWELVE_HOURS ? "h12" : "h23"
 }
