+++ test/tests/settings/UserDataExportTest.ts
@@ -5,7 +5,7 @@ import { LoginController } from "../../../src/api/main/LoginController.js"
 import { FileController } from "../../../src/file/FileController.js"
 import { object, when } from "testdouble"
 import { CustomerTypeRef, Group, GroupInfo, GroupInfoTypeRef, GroupTypeRef, User } from "../../../src/api/entities/sys/TypeRefs.js"
-import { formatDateTimeUTC } from "../../../src/calendar/export/CalendarImporter.js"
+import { formatDateTimeUTC } from "../../../src/calendar/export/CalendarExporter.js"
 import { CounterFacade } from "../../../src/api/worker/facades/lazy/CounterFacade.js"
 import { CounterType } from "../../../src/api/common/TutanotaConstants.js"
 import { CounterValueTypeRef } from "../../../src/api/entities/monitor/TypeRefs.js"
